backend/requirements.txt
# Django
Django==5.0.6
djangorestframework==3.15.1

# Linting
black==25.1.0
flake8==7.2.0
isort==6.0.0

# Testing
pytest==8.3.4
pytest-django==4.15.0
pytest-cov==6.0.0
pytest-mock==3.14.0
